\name{landscape.new.example}
\alias{landscape.new.example}
\title{Create a Default Landscape}

\description{
  Create a Rmetasim landscape with all default parameters.}
\usage{
  rland <- landscape.new.example()
}
\arguments{
  None
}
\examples{
  ## Only usage
  landscape.new.example()}
\keyword{misc}
